+++ config.sample.php
@@ -3,9 +3,9 @@
 
 // database connection
 define('DATABASE_HOSTNAME', "localhost");
-define('DATABASE_USERNAME', "");
-define('DATABASE_PASSWORD', "");
-define('DATABASE_NAME', "");
+define('DATABASE_USERNAME', '');
+define('DATABASE_PASSWORD', '');
+define('DATABASE_NAME', '');
 
 // Mail
 define('MAIL_HOST', '');
@@ -14,18 +14,11 @@ define('MAIL_PASSWORD', '');
 define('MAIL_FROM', '');
 define('MAIL_TO', '');
 
-// Facebook's Graph API
-define('FACEBOOK_API_ID', "");
-define('FACEBOOK_API_SECRET', "");
-
-// Google API
-define('GOOGLE_API_KEY', "");
-
 // Twitter API
-define('TWITTER_API_TOKEN', "");
-define('TWITTER_API_TOKEN_SECRET', "");
-define('TWITTER_API_CONSUMER_KEY', "");
-define('TWITTER_API_CONSUMER_SECRET', "");
+define('TWITTER_API_TOKEN', '');
+define('TWITTER_API_TOKEN_SECRET', '');
+define('TWITTER_API_CONSUMER_KEY', '');
+define('TWITTER_API_CONSUMER_SECRET', '');
 
 // OpenCage API
 define('OPENCAGE_API_KEY', '');