+++ CHANGELOG.md
@@ -5,10 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Version 3.6.3 (unreleased)
+
+## Version 3.6.3 (2019-12-19)
 
 ### Fixed
 * Map, Search, Details, Add Form: Removed Google+
+* Fix update of Facebook follower count
+* Comparison: Removed YouTube (see [#31](https://github.com/patrickrobrecht/kirchen-im-web/issues/31))
 
 
 ## Version 3.6.2 (2019-09-06)